// TakosEncryptInk/types.ts
type Sign = {
  signature: string // ArrayBufferをbase64に変換したもの
  hashedPublicKeyHex: string // 公開鍵をハッシュ化し、16進数文字列に変換したもの
  version: number // 署名のバージョン
}
// ...
interface KeyShareSignKeyPub {
  key: string
  timestamp: string
  sign: Sign
  keyType: "keyShareSignPub" // 鍵の種類
  version: number // 鍵のバージョン
}

interface KeyShareSignKeyPrivate {
  key: string
  keyType: "keyShareSignPrivate" // 鍵の種類
  version: number // 鍵のバージョン
  timestamp: string
}

interface KeyShareSignKey {
  public: KeyShareSignKeyPub
  private: KeyShareSignKeyPrivate
  hashHex: string
  version: number
}

interface migrateKeyPub {
  key: string
  keyType: "migratePub"
  version: number
}

interface migrateKeyPrivate {
  key: string
  keyType: "migratePrivate"
  version: number
}

interface migrateKey {
  public: migrateKeyPub
  private: migrateKeyPrivate
  hashHex: string
  version: number
}

interface migrateDataSignKeyPub {
  key: string
  keyType: "migrateDataSignPub"
  version: number
}

interface migrateDataSignKeyPrivate {
  key: string
  keyType: "migrateDataSignPrivate"
  version: number
}

interface migrateDataSignKey {
  public: migrateDataSignKeyPub
  private: migrateDataSignKeyPrivate
  hashHex: string
  version: number
}